wireless: Support can-scan-one logic.
Enable this by passing a -1 for a scan frequency.

When enabled, the system will only scan the current active
channel if at least one VIF is actively using it.  If no
VIFS are active or this flag is disabled, then default
behaviour is used.

This helps when using multiple STA interfaces that otherwise might
constantly be trying to scan all channels.

Signed-off-by: Ben Greear <greearb@candelatech.com>