ath10k:  Add firmware feature flag for rx-sw-crypt support.
The 'community' version of the CT firmware will not support
rx-sw-crypt (which enables encrypted access for multiple
station vifs to the same AP).  Tweak the multi-vif support
to deal with this and allow multi-vif-to-same AP without
encryption on the community version.  The commercial
CT firmware version will do full rx-sw-crypt support.

Signed-off-by: Ben Greear <greearb@candelatech.com>